API: cleanup
